A lot of changes

+++ evaluate_models.py
@@ -18,8 +18,6 @@ use datafra,e
 #from nbformat import write
 from decision_mappings import ImageNetProbabilitiesTo1000ClassNamesMapping, ImageNetProbabilitiesTo16ClassNamesMapping
 from helper import wordnet_functions as wnf
-
- 
 from torchvision.datasets import ImageFolder
 from torch.utils.data import DataLoader
 import torchvision.models as models
@@ -29,19 +27,20 @@ import torch
 from cue_conflict_dataset import CueConflictDataloader
 import csv
 from tqdm import tqdm
-from pl_bolts.models.self_supervised import SimCLR
 import os
-from sklearn.neighbors import KNeighborsClassifier
-
+import numpy as np
+import copy
+import pandas as pd
+import re
 
 
 class ShapeBiasEvaluator:
-    def __init__(self,dataset_path,resize:bool=True,batch_size=128,workers=4,device=None):
+    def __init__(self,dataset_path,resize:bool=False,batch_size=128,workers=4,device=None):
         self.dataset_path = dataset_path
         self.resize = resize
         self.batch_size = batch_size
         self.workers = workers
-        self.dataloader = CueConflictDataloader()("cue-conflict",True,self.batch_size,self.workers)
+        self.dataloader = CueConflictDataloader()("cue-conflict",resize,self.batch_size,self.workers)
         if device is None:
             self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
         else:
@@ -51,47 +50,39 @@ class ShapeBiasEvaluator:
     def _run(self,model,filename="results.csv",max_lines=None):
         max_lines = max_lines if max_lines is not None else -1        
         mapp = ImageNetProbabilitiesTo16ClassNamesMapping()
+        paths, logits = [], []
         with open(filename,"w") as fp:
             writer = csv.writer(fp)
             writer.writerow(["response", "shape","texture", "imagename"])
-            for x,shapes,textures,paths in tqdm(self.dataloader):
+            for x, shapes, textures, path_batch in tqdm(self.dataloader):
                 with torch.no_grad():
-                    x = x.to(self.device)
-                    output = torch.nn.functional.softmax(model(x),dim=1)
-                    classes_batch = mapp(output.cpu().numpy())
-                    for classes,shape,texture,path in zip(classes_batch,shapes,textures,paths):
-                        writer.writerow([classes[0],shape,texture,path])
-                        max_lines -= 1
-                        if max_lines == 0: break
-                if max_lines == 0: break
+                    output = model(x.to(self.device))
+                logits.append(output)
+                paths.append(path_batch)
+            logits = torch.stack(logits)
+            probabilities = torch.nn.functional.softmax(logits,dim=1)
+            label_choices = mapp(probabilities.cpu().numpy())
+            for classes,shape, texture, path in zip(label_choices, shapes, textures, paths):
+                writer.writerow([classes[0],shape,texture,path])
 
-    def _knn(self,model,X,raw_y,filename="results.csv",max_lines=None):
-        mapp = ImageNetProbabilitiesTo16ClassNamesMapping()
+        
 
-        a = torch.zeros(1,1000)
-        a[raw_y] = 1.0
-        classes_batch = mapp(a.cpu().numpy())
-
-        print(classes_batch)
-
-        '''
-        max_lines = max_lines if max_lines is not None else -1        
-        mapp = ImageNetProbabilitiesTo16ClassNamesMapping()
-        with open(filename,"w") as fp:
-            writer = csv.writer(fp)
-            writer.writerow(["response", "shape","texture", "imagename"])
-            for x,shapes,textures,paths in tqdm(self.dataloader):
-                with torch.no_grad():
-                    x = x.to(self.device)
-                    features = model(x)
-                
-                    classes_batch = mapp(output.cpu().numpy())
-                    for classes,shape,texture,path in zip(classes_batch,shapes,textures,paths):
-                        writer.writerow([classes[0],shape,texture,path])
-                        max_lines -= 1
-                        if max_lines == 0: break
-                if max_lines == 0: break
-        '''
+        ## WRITE CSV
+        #total_logits = np.concatenate(total_logits) #!
+        #softmax = torch.softmax(torch.tensor(total_logits),dim=1)
+        #classes_ = torch.argmax(softmax,dim=1)
+        """
+        print(total_logits.shape) #!
+        l = pd.DataFrame(total_logits) #!
+        p = pd.DataFrame(total_paths,columns={"filename"}) #!
+        p["filename"] = p["filename"].apply(lambda x: re.sub(".*/.*/","",x))
+        l.join(p).to_csv("logits.csv") #!
+        print(len(total_paths)) #!
+        print(classes_)
+        pd.DataFrame(softmax.numpy()).join(p).to_csv("original_softmax.csv")
+        pd.DataFrame(classes_.numpy()).join(p).to_csv("original_1000classes.csv")
+        """
+        
 
     def _summarize(self,filename="results.csv",verbose=True):
         results = {}
@@ -121,10 +112,12 @@ class ShapeBiasEvaluator:
 
     def evaluate(self,model,filename="results.csv",max_lines=None,verbose=True,keep_result_file=False):
         model.to(self.device)
+        model.eval()
         self._run(model,filename=filename)
         results, info = self._summarize(filename,verbose)
         assert os.path.exists(filename)
-        os.remove(filename)
+        if keep_result_file == False:
+            os.remove(filename)
         return results, info
 
     
@@ -134,14 +127,15 @@ class ShapeBiasEvaluator:
 if __name__ == '__main__':
     import timm
     import sys
+    model = models.alexnet(pretrained=True)
     #model = models.resnet18(pretrained=True)
-    model = timm.create_model('resnet18d',pretrained=True)
-    dataloader = CueConflictDataloader()("cue-conflict",True,128,4)
+    #model = timm.create_model('resnet18d',pretrained=True)
+    #dataloader = CueConflictDataloader()("cue-conflict",True,128,4)
     file_name = "results.csv"
     max_lines = -1
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     verbose = True
-    ShapeBiasEvaluator("cue-conflict").evaluate(model)
+    ShapeBiasEvaluator("cue-conflict").evaluate(model,keep_result_file=True)
